fix(cli): move feedback misses out of app package
Keep closed-loop feedback code under the CLI package so the repo no longer ships a stray app namespace.
