+++ tests/cli/test_misses_command.py
@@ -8,8 +8,8 @@ from pathlib import Path
 import pytest
 from click.testing import CliRunner
 
-from app.feedback import MissTaxonomy, record_miss
 from cli.__main__ import cli
+from cli.feedback import MissTaxonomy, record_miss
 
 
 @pytest.fixture
